[lint package refactor] Create a file for the report functions
